Backend/requirements.txt
flask
flask-cors
mcp
uvicorn
httpx
python-dotenv
pandas
scikit-learn
numpy
PyPDF2
sentence-transformers
ctransformers>=0.2.27
fastapi
sqlalchemy
pymysql
transformers
torch
accelerate
bitsandbytes
# Chat History Dependencies
aiomysql
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
